requirements.txt
Pillow
flask
scikit-learn
open_clip_torch
torch
torchvision